+++ audit_tables.gemspec
@@ -33,10 +33,11 @@ Gem::Specification.new do |spec|
   spec.require_paths = ['lib']
 
   spec.add_development_dependency 'bundler', '~> 1.13'
+  spec.add_development_dependency 'byebug', '~> 9.0.6'
+  spec.add_development_dependency 'faker'
+  spec.add_development_dependency 'pg'
   spec.add_development_dependency 'rake', '~> 10.0'
   spec.add_development_dependency 'rails'
   spec.add_development_dependency 'rspec', '~> 3.0'
   spec.add_development_dependency 'rubocop', '= 0.42.0'
-  spec.add_development_dependency 'pg'
-  spec.add_development_dependency 'faker'
 end
