+++ lib/audit_tables/base.rb
