+++ lib/audit_tables/build_audit_trigger.rb
