+++ lib/audit_tables/create_audit_tables_for_existing_tables.rb
