+++ lib/audit_tables/create_new_audit_table.rb
